Чистка и рефакторинг
На локалке все работает, надеюсь пути не поломаются на гите.

-Убрал все лишние файлы,
-Переименовал нужные чуть более понятно
-Сделал рефакторинг модалки, библиотеки и сторожа
-Вынес функции в отдельные файлы, добавил где нужно импорты

Еще нужно будет пересмотреть, есть еще что почистить
