+++ src/index.js
@@ -1,11 +1,6 @@
 import './js/apiService';
 import './js/search';
-import storage from './js/storage/modal-home';
-import storageGetQueue from './js/storage/get-queue';
-import storageAddWatched from './js/storage/add-watched';
-import btnQueue from './js/storage/modal-library-queue';
-import btnQueue from './js/storage/modal-library-watched';
 import './js/library/library-filters';
 import './js/modal/modal';
 import './js/pagination';
-import './js/homeToLib';
+import './js/nav-btns-classes';
