+++ src/js/library/library-filters.js
@@ -1,49 +1,21 @@
 import { refs } from '../refs';
-import { renderMoviesDetailed } from './library-render-from-local-storage';
+import { renderMoviesDetailed } from './initial-library-render';
+import { toggleActiveClass } from './fnToggleActiveClass';
+import { getFromLocalStorage } from '../storage/ls-data-services';
+import { renderMoviesDetailed } from './fnrenderMoviesDetailed';
 
 refs.libButtons.addEventListener('click', onLibraryBtnClick);
-
+// renderMoviesDetailed - Фильтрация рендера фильмов по нажатию кнопки Watched или Queue
+// toggleActiveClass - Добавляет класс active для и соответсвующий data-tag
 async function onLibraryBtnClick(event) {
   if (event.target.nodeName !== 'BUTTON') {
     return;
   }
   const currentActiveTarget = refs.libButtons.querySelector('.active');
   const selectedTarget = event.target;
-
-  await toggleActiveClass(currentActiveTarget, selectedTarget);
   const moviesFromStorage = getFromLocalStorage(
     selectedTarget.textContent.toLowerCase()
   );
+  await toggleActiveClass(currentActiveTarget, selectedTarget);
   renderMoviesDetailed(moviesFromStorage);
 }
-
-function toggleActiveClass(currentActiveTarget, selectedTarget) {
-  if (currentActiveTarget) {
-    currentActiveTarget.removeAttribute('data-tag');
-    currentActiveTarget.classList.remove('active');
-  }
-  selectedTarget.classList.add('active');
-  selectedTarget.setAttribute(
-    'data-tag',
-    selectedTarget.innerText.trim().toLowerCase()
-  );
-}
-
-function getFromLocalStorage(tag) {
-  let storedMovies = JSON.parse(localStorage.getItem(tag)) || [];
-  return storedMovies;
-}
-
-refs.libraryBtn.addEventListener('click', getAllMoviesFromLocalStorage);
-
-function getAllMoviesFromLocalStorage() {
-  const watchedMovies = getFromLocalStorage('watched');
-  const queueMovies = getFromLocalStorage('queue');
-  const allMovies = [...watchedMovies, ...queueMovies];
-  const uniqueMovies = [...new Set(allMovies.map(movie => movie.id))].map(id =>
-    allMovies.find(movie => movie.id === id)
-  );
-  return uniqueMovies;
-}
-
-export { getAllMoviesFromLocalStorage };
