+++ src/js/modal/modal.js
@@ -6,12 +6,15 @@ import {
   closeModalOnEsc,
   closeModalOnBtnClick,
 } from './modal-close-functions';
+import { onAddToQueueClick, onAddToWatchedClick } from './modal-btns-functions';
 
 const api = new MovieApi();
-refs.gallery.addEventListener('click', async e => {
+refs.gallery.addEventListener('click', onGalleryCardClick);
+
+async function onGalleryCardClick(e) {
   let targetId = e.target.closest('li').attributes[0].value;
-  await api.getMovieDetails(targetId).then(movie => {
-    refs.modalContent.innerHTML = modalCard(movie, targetId);
+  api.getMovieDetails(targetId).then(movie => {
+    refs.modalContent.innerHTML = modalCard(movie);
     refs.modal.style.display = 'block';
     document.body.style.overflow = 'hidden';
 
@@ -20,52 +23,9 @@ refs.gallery.addEventListener('click', async e => {
         onAddToWatchedClick(movie);
       } else if (e.target.matches('.addQueue')) {
         onAddToQueueClick(movie);
-      } else {
-        return;
       }
     });
   });
-});
-
-function addToLocalStorage(tag, movieData, key) {
-  let movies = JSON.parse(localStorage.getItem(tag)) || [];
-
-  if (!localStorage.getItem(key)) {
-    localStorage.setItem(key, JSON.stringify([]));
-  }
-
-  if (tag === 'watched') {
-    const queueMovies = JSON.parse(localStorage.getItem('queue')) || [];
-    const movieIndex = movies.findIndex(movie => movie.id === movieData.id);
-    if (movieIndex === -1) {
-      movies.push(movieData);
-      localStorage.setItem(tag, JSON.stringify(movies));
-    } else {
-      console.log('Movie is already in the array');
-    }
-  }
-
-  if (tag === 'queue') {
-    const watchedMovies = JSON.parse(localStorage.getItem('watched')) || [];
-    const movieIndex = movies.findIndex(movie => movie.id === movieData.id);
-    const watchedMovieIndex = watchedMovies.findIndex(
-      movie => movie.id === movieData.id
-    );
-    if (movieIndex === -1 && watchedMovieIndex === -1) {
-      movies.push(movieData);
-      localStorage.setItem(tag, JSON.stringify(movies));
-    } else {
-      console.log('Movie is already in the array');
-    }
-  }
-}
-
-function onAddToWatchedClick(movieData) {
-  addToLocalStorage('watched', movieData, 'watched');
-}
-
-function onAddToQueueClick(movieData) {
-  addToLocalStorage('queue', movieData, 'queue');
 }
 
 refs.closeBtn.addEventListener('click', closeModalOnBtnClick);
