--- src/js/storage/add-queue.js
@@ -1,11 +1,0 @@
-function addQueue(data) {
-    
-    try {
-        localStorage.setItem('films queue', JSON.stringify(data))
-    } catch (error) {
-        console.log('storage error', error.name)
-        console.log(error.message)
-    }
-};
-
-export default { addQueue };
