--- src/js/storage/add-watched.js
@@ -1,11 +1,0 @@
-function addWatched(data) {
-    
-    try {
-        localStorage.setItem('films watched', JSON.stringify(data))
-    } catch (error) {
-        console.log('storage error', error.name)
-        console.log(error.message)
-    }
-}
-
-export default { addWatched };
