--- src/js/storage/get-queue.js
@@ -1,13 +1,0 @@
-function getQueue() {
-    
-    try {
-        let jsonQueue = localStorage.getItem('films queue');
-        let dataQueue = JSON.parse(jsonQueue)
-        return dataQueue
-    } catch (error) {
-        console.log('storage error', error.name)
-        console.log(error.message)
-    }
-};
-
-export default { getQueue };
