--- src/js/storage/lock-modal-queue.js
@@ -1,22 +1,0 @@
-import storageGetQueue from './get-queue';
-
-function lockModalQueue(card) {
-    let idFilm = card.id;
-    let lockQueue = false;
-    let storageQueue = [];
-
-    storageQueue = storageGetQueue.getQueue();
-    
-    if (storageQueue) {
-        if (!storageQueue.find(item => item.id === idFilm)) {
-            lockQueue = true;
-        }
-    } else {
-        storageQueue = [];
-        lockQueue = true
-    }
-    
-    return {lockQueue, storageQueue}
-}
-
-export default {lockModalQueue}
