--- src/library.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="ru">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Your Library</title>
-    <link rel="stylesheet" href="./sass/index.scss" />
-    <script type="module" src="./index.js"></script>
-  </head>
-  <body>
-    <!-- Шапка должна отличаться от главной страницы (Паша) -->
-    <!-- СЮда пойдет импорт разметки библиотеки фильмов (Оля) -->
-    <include src="./partials/footer.html"></include>
-  </body>
-</html>
