--- src/partials/cards.html
@@ -1,12 +1,0 @@
-<li class="gallery__item" data-id="">
-            
-            <img class="gallery__item-image" src=""     
-            alt=""
-            width="280"
-            loading="lazy"
-            />
-            <div class="gallery__item-description">
-            <p class="gallery__item-description-title"> title </p>
-            <p class="gallery__item-description-genres"> genre | year </p>
-            </div>
-</li>
